converted all bytes to ints except for enums, made several style and readability improvements (new enums), and added the first four Gen III moves.

+++ Assets/Scripts/constants/PokemonConst.cs
@@ -1,4 +1,4 @@
 public static class PokemonConst
 {
-    public const byte maxLevel = 100;
+    public const int maxLevel = 100;
 }
