+++ Assets/Scripts/data/include/WildDataset.cs
@@ -19,14 +19,14 @@
         var random = new System.Random();
         int i = random.Next() % TotalFreq;
         SpeciesID species = SpeciesID.Missingno;
-        byte level = 10;
+        int level = 10;
         foreach (WildMon mon in wildMons)
         {
             i -= mon.freq;
             if (i < 0)
             {
                 species = mon.species;
-                level = (byte)(mon.minLevel + (random.Next() % (mon.maxLevel - mon.minLevel + 1)));
+                level = (mon.minLevel + (random.Next() % (mon.maxLevel - mon.minLevel + 1)));
                 break;
             }
         }
