+++ Assets/Scripts/data/teamData/include/EvIvStruct.cs
@@ -3,16 +3,16 @@
 public struct EvIvSpread
 {
     public bool real;
-    public byte ivHP;
-    public byte ivAttack;
-    public byte ivDefense;
-    public byte ivSpAtk;
-    public byte ivSpDef;
-    public byte ivSpeed;
-    public byte evHP;
-    public byte evAttack;
-    public byte evDefense;
-    public byte evSpAtk;
-    public byte evSpDef;
-    public byte evSpeed;
+    public int ivHP;
+    public int ivAttack;
+    public int ivDefense;
+    public int ivSpAtk;
+    public int ivSpDef;
+    public int ivSpeed;
+    public int evHP;
+    public int evAttack;
+    public int evDefense;
+    public int evSpAtk;
+    public int evSpDef;
+    public int evSpeed;
 }
